Major improvements: fix compilation errors, add security hardening, fix broken I/O
Compilation Fixes (8 files):
- hal/cpu/syscall.h: Renumber SYS_USER_* syscalls (0-7 → 120-127) to fix
  duplicate case value conflicts with SYS_EXIT/SYS_EXEC/SYS_EXEC_ARGS/SYS_GET_ARGS
- hal/video/compositor_v2.c: Add missing #include for strlen()
- hal/video/cgcontext.h: Add typedef for CGContext/CGFont/CGImage structs
  so sizeof() works in C (not just C++)
- usr/desktop.c: Add forward declaration for desktop_apply_selection()
- usr/apps/disk_utility.c: Add missing #include for s_printf
- usr/apps/image_viewer.c: Add missing #include for s_printf
- usr/libs/js_engine_v2.c: Fix _js_memset/_js_memcpy/_js_memmove type
  mismatch (unsigned long → size_t)
- usr/libs/browser_enhanced.c: Add static inline strncasecmp() implementation
- usr/libs/browser_js_bridge.c: Include browser_dom.h for full dom_node_t
  definition, fix field names (tag_name→tag, type==1→DOM_NODE_ELEMENT)

Security Hardening:
- NEW: core/user_copy.c/h - Implement copy_from_user/copy_to_user with
  validate_user_ptr() that checks against KERNEL_BASE (0xC0000000) to
  prevent Ring 3 processes from reading/writing kernel memory via syscalls
- NEW: core/spinlock.c/h - Implement kernel spinlocks with xchg-based
  atomic acquire, memory barrier release, and IRQ-safe variants
  (spinlock_irqsave_acquire/release). Added g_heap_lock to kmalloc/kfree
- core/tls.c: Implement actual X.509 certificate verification including
  expiry checking, hostname verification (CN + SAN with wildcard support),
  RSA signature verification via SHA-256, and CA trust store lookup.
  Self-signed certs not in trust store are now REJECTED instead of accepted.
- hal/cpu/syscall.c: Add user pointer validation to all SYS_USER_* syscalls

Broken I/O Fixes:
- core/bsd_syscall.c: Fix bsd_read() to use handle-based offset I/O
  instead of reading entire file into 4KB stack buffer. Fix bsd_write()
  to write at offset instead of replacing entire file contents. Track
  file offsets in fd entries.
- hal/cpu/syscall.c: Wire SYS_FORK to process_fork() and implement
  basic SYS_WAITPID (zombie reaping)

Build System Improvements:
- Makefile: Add -MMD -MP for auto-dependency generation
- Makefile: Protect disk.img from accidental recreation
- Makefile: Add core/user_copy.c and core/spinlock.c to build
- Makefile: Add check-nasm prerequisite for assembly targets
- Makefile: Add distclean target, improve clean to remove .d files

All 126 C source files compile with zero errors.

+++ usr/libs/js_engine_v2.c
@@ -81,20 +81,20 @@ static inline char* _js_strstr(const char* haystack, const char* needle) {
     return 0;
 }
 
-static inline void* _js_memset(void* s, int c, unsigned long n) {
+static inline void* _js_memset(void* s, int c, size_t n) {
     unsigned char* p = (unsigned char*)s;
     while (n--) *p++ = (unsigned char)c;
     return s;
 }
 
-static inline void* _js_memcpy(void* dest, const void* src, unsigned long n) {
+static inline void* _js_memcpy(void* dest, const void* src, size_t n) {
     unsigned char* d = (unsigned char*)dest;
     const unsigned char* s = (const unsigned char*)src;
     while (n--) *d++ = *s++;
     return dest;
 }
 
-static inline void* _js_memmove(void* dest, const void* src, unsigned long n) {
+static inline void* _js_memmove(void* dest, const void* src, size_t n) {
     unsigned char* d = (unsigned char*)dest;
     const unsigned char* s = (const unsigned char*)src;
     if (d < s) while (n--) *d++ = *s++;
